Write the client and ctrl tests ahead of the split
client.test.ts drives ./client action-first with kebab-case flags against a
stub proxy; ctrl.test.ts drives ./ctrl the same way. Linear and createExperiment
unit tests move next to the modules they will exercise, and prompt() takes the
Cursor api key as an argument.
